chore(router): Minor reorganization
